+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Add 16 bit dataframe size for `SPI`
+- Implement `timer::Cancel` trait for `CountDownTimer`
 
 ### Added